# openapi/v1.yaml
openapi: 3.1.0
info:
  title: Forge local control API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
  description: |
    Local-only HTTP/JSON and SSE control contract for Forge.

    This API is served over an owner-only Unix domain socket, not a TCP
    listener. The nominal server URL only supplies an HTTP authority to generic
    OpenAPI clients. M0 has no application-level authentication: the local
    process is the sole operator. Every mutation is a named command, committed
    by Core, and requires an `Idempotency-Key` header plus an expected Project
    revision. The operator is recorded in canonical Events as a `human` actor;
    M0 defines no separate local-administrator actor kind.
servers:
  - url: http://forge.local
    description: Nominal HTTP authority carried over the local Unix socket.
...
security: []
...
paths:
...
  /v1/projects/{project_id}/hook-versions:
    get:
      tags: [projects]
      operationId: listProjectHookVersions
      summary: Read immutable owner-configured command versions; never execute them.
      parameters:
        - {name: project_id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: after, in: query, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
      responses:
        '200':
          description: Ascending version-ID page. Names may repeat; Pipeline pins exact IDs.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: {type: array, items: {$ref: '#/components/schemas/ProjectHookVersionView'}}
                  next_cursor: {$ref: '#/components/schemas/UuidV7'}
        '400': {description: Invalid identifier, cursor or page limit.}
        '404': {description: Project absent.}
  /v1/projects/{project_id}/hook-invocations:
    get:
      tags: [projects]
      operationId: listProjectHookInvocations
      summary: Read bounded canonical hook status and mapped result metadata; private runtime output is excluded.
      parameters:
        - {name: project_id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: after, in: query, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses:
        '200':
          description: Ascending invocation-ID page for the requested Project.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/ProjectHookInvocationView'}}
                  next_cursor: {$ref: '#/components/schemas/UuidV7'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
  /v1/projects/{projectId}/runs/{runId}/context:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/RunId'
    get:
      tags: [runs]
      operationId: getRunContextCoordinates
      summary: Read safe dispatch coordinates without the context snapshot body.
      responses:
        '200':
          description: Available coordinates or an explicit unavailable marker.
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    additionalProperties: false
                    required: [availability, coordinates]
                    properties:
                      availability: {type: string, const: available}
                      coordinates:
                        type: object
                        additionalProperties: false
                        required: [context_snapshot_id, run_id, project_id, task_id, employee_id, pipeline_version_id, stage_id, stage_visit, task_revision_before_dispatch, run_spec_id]
                        properties:
                          context_snapshot_id: {$ref: '#/components/schemas/UuidV7'}
                          run_id: {$ref: '#/components/schemas/UuidV7'}
                          project_id: {$ref: '#/components/schemas/UuidV7'}
                          task_id: {$ref: '#/components/schemas/UuidV7'}
                          employee_id: {$ref: '#/components/schemas/UuidV7'}
                          pipeline_version_id: {$ref: '#/components/schemas/UuidV7'}
                          stage_id: {type: string}
                          stage_visit: {type: [integer, 'null'], minimum: 1}
                          task_revision_before_dispatch: {type: integer, minimum: 1}
                          run_spec_id: {$ref: '#/components/schemas/UuidV7'}
                  - type: object
                    additionalProperties: false
                    required: [availability, run_id, coordinates]
                    properties:
                      availability: {type: string, const: unavailable}
                      run_id: {$ref: '#/components/schemas/UuidV7'}
                      coordinates: {type: 'null'}
        '404': {$ref: '#/components/responses/NotFound'}
  /v1/projects/{projectId}/runs/{runId}/evidence:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/RunId'
      - name: cursor
        in: query
        schema: {$ref: '#/components/schemas/UuidV7'}
      - name: limit
        in: query
        schema: {type: integer, minimum: 1, maximum: 50, default: 20}
    get:
      tags: [runs]
      operationId: listRunEvidenceReceipts
      summary: Read bounded metadata receipts without evidence content or storage keys.
      responses:
        '200':
          description: One page of safe receipts.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    maxItems: 50
                    items: {$ref: '#/components/schemas/RunEvidenceReceipt'}
                  next_cursor: {$ref: '#/components/schemas/UuidV7'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UuidV7'
...
    RunId:
      name: runId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UuidV7'
...
  responses:
    InvalidRequest:
      description: The request cannot be parsed or violates a transport contract.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: The requested resource does not exist in the Project scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Unavailable:
      description: A required local Forge dependency is temporarily unavailable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    UuidV7:
      type: string
      format: uuid
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
      description: A UUIDv7 identifier generated by Forge unless stated otherwise.
    Rfc3339Timestamp:
      type: string
      format: date-time
      description: UTC timestamp serialized in RFC 3339 format.
    JsonObject:
      type: object
      additionalProperties: true
      description: A JSON object interpreted only by the selected domain contract.
...
    ProjectHookVersionView:
      type: object
      additionalProperties: false
      required: [id, project_id, name, image, command, workdir, limits, max_output_bytes, applicable_task_kinds, required, created_at]
      properties:
        id: {$ref: '#/components/schemas/UuidV7'}
        project_id: {$ref: '#/components/schemas/UuidV7'}
        name: {$ref: '#/components/schemas/ConfigureProjectHookPayload/properties/name'}
        image: {$ref: '#/components/schemas/ConfigureProjectHookPayload/properties/image'}
        command: {$ref: '#/components/schemas/ConfigureProjectHookPayload/properties/command'}
        workdir: {$ref: '#/components/schemas/ConfigureProjectHookPayload/properties/workdir'}
        limits: {$ref: '#/components/schemas/RuntimeBinding/properties/limits'}
        max_output_bytes: {$ref: '#/components/schemas/ConfigureProjectHookPayload/properties/max_output_bytes'}
        applicable_task_kinds: {$ref: '#/components/schemas/ConfigureProjectHookPayload/properties/applicable_task_kinds'}
        required: {type: boolean}
        created_at: {type: string, format: date-time}
    ProjectHookInvocationView:
      type: object
      additionalProperties: false
      required: [id, task_id, pipeline_version_id, stage_id, stage_visit, hook_version_id, candidate_proposal_id, run_id, state, verdict, mapped_outcome, artifact_id, created_at, updated_at]
      properties:
        id: {$ref: '#/components/schemas/UuidV7'}
        task_id: {$ref: '#/components/schemas/UuidV7'}
        pipeline_version_id: {$ref: '#/components/schemas/UuidV7'}
        stage_id: {type: string, minLength: 1, maxLength: 128}
        stage_visit: {type: integer, minimum: 1}
        hook_version_id: {$ref: '#/components/schemas/UuidV7'}
        candidate_proposal_id: {type: [string, 'null'], format: uuid}
        run_id: {type: [string, 'null'], format: uuid}
        state: {type: string, enum: [running, completed, held]}
        verdict: {type: [string, 'null'], enum: [passed, failed, timed_out, skipped, null]}
        mapped_outcome: {type: [string, 'null'], maxLength: 128}
        artifact_id: {type: [string, 'null'], format: uuid}
        created_at: {$ref: '#/components/schemas/Rfc3339Timestamp'}
        updated_at: {$ref: '#/components/schemas/Rfc3339Timestamp'}
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - invalid_request
            - validation_failed
            - not_found
            - conflict
            - stale_revision
            - idempotency_conflict
            - forbidden
            - execution_stopped
            - cursor_invalid
            - unavailable
            - internal
        message:
          type: string
        details:
          $ref: '#/components/schemas/JsonObject'
        request_id:
          $ref: '#/components/schemas/UuidV7'
...
    RunEvidenceReceipt:
      type: object
      additionalProperties: false
      required: [id, run_id, task_id, stream, sequence, sha256, size_bytes, redaction_policy_reference, storage_state, created_at, content_availability]
      properties:
        id: {$ref: '#/components/schemas/UuidV7'}
        run_id: {$ref: '#/components/schemas/UuidV7'}
        task_id:
          oneOf:
            - {$ref: '#/components/schemas/UuidV7'}
            - {type: 'null'}
        stream: {type: string, enum: [stdout, stderr, diagnostic]}
        sequence: {type: integer, minimum: 1}
        sha256: {type: string, pattern: '^[0-9a-f]{64}$'}
        size_bytes: {type: integer, minimum: 1}
        redaction_policy_reference: {type: string, minLength: 1}
        storage_state: {type: string, enum: [pending_upload, stored, unconfirmed]}
        created_at: {$ref: '#/components/schemas/Rfc3339Timestamp'}
        content_availability: {type: string, const: unavailable}
...